Internal

Internal API

Types

RCall.CharSxp — Type.

R character string

source

RCall.ClosSxp — Type.

R function closure

source

RCall.CplxSxp — Type.

R complex vector

source

RCall.EnvSxp — Type.

R environment

source

RCall.IntSxp — Type.

R integer vector

source

RCall.LangSxp — Type.

R function call

source

RCall.LglSxp — Type.

R logical vector

source

RCall.ListSxp — Type.

R pairs (cons) list cell

source

RCall.NilSxp — Type.

R NULL value

source

RCall.RObject — Type.

An RObject is a Julia wrapper for an R object (known as an "S-expression" or "SEXP"). It is stored as a pointer which is protected from the R garbage collector, until the RObject itself is finalized by Julia. The parameter is the type of the S-expression.

When called with a Julia object as an argument, a corresponding R object is constructed.

julia> RObject(1)
RObject{IntSxp}
[1] 1

julia> RObject(1:3)
RObject{IntSxp}
[1] 1 2 3

julia> RObject(1.0:3.0)
RObject{RealSxp}
[1] 1 2 3

source

RCall.RObject — Method.

sexp(x) converts a Julia object x to a pointer to a corresponding Sxp Object.

source

RCall.RealSxp — Type.

R real vector

source

RCall.S4Sxp — Type.

R S4 object

source

RCall.StrSxp — Type.

R vector of character strings

source

RCall.Sxp — Type.

RCall.jl's type Sxp mirrors the R symbolic expression record SEXPREC in R API. These are represented by a pointer Ptr{S<:Sxp} (which is called SEXP in R API).

source

RCall.VecSxp — Type.

R list (i.e. Array{Any,1})

source

RCall.AnySxp — Type.

R "any" object

source

RCall.BcodeSxp — Type.

R byte code

source

RCall.BuiltinSxp — Type.

R built-in function

source

RCall.DotSxp — Type.

R dot-dot-dot object

source

RCall.ExprSxp — Type.

R expression vector

source

RCall.ExtPtrSxp — Type.

R external pointer

source

RCall.PromSxp — Type.

R promise

source

RCall.RawSxp — Type.

R byte vector

source

RCall.SpecialSxp — Type.

R special function

source

RCall.SxpHead — Type.

R Sxp header: a pointer to this is used for unknown types.

source

RCall.SymSxp — Type.

R symbol

source

RCall.WeakRefSxp — Type.

R weak reference

source

Methods

Modules = [RCall, RCall.Console, RCall.IJuliaHooks]
Order   = [:function]

Macros

RCall.@R_str — Macro.

R"..."

An inline R expression, the result of which is evaluated and returned as an RObject.

It supports substitution of Julia variables and expressions via prefix with $ whenever not valid R syntax (i.e. when not immediately following another completed R expression):

R"glm(Sepal.Length ~ Sepal.Width, data=$iris)"

It is also possible to pass Julia expressions:

R"plot(RCall.#86)"

All such Julia expressions are evaluated once, before the R expression is evaluated.

The expression does not support assigning to Julia variables, so the only way retrieve values from R via the return value.

source

RCall.@rget — Macro.

Copies variables from R to Julia using the same name.

source

RCall.@rimport — Macro.

Import an R Package as a Julia module. For example,

@rimport ggplot2

is equivalent to ggplot2 = rimport("ggplot2") with error checking.

You can also use classic Python syntax to make an alias: @rimport *package-name* as *shorthand*

@rimport ggplot2 as gg

which is equivalent to gg = rimport("ggplot2").

source

RCall.@rlibrary — Macro.

Load all exported functions/objects of an R package to the current module. Almost equivalent to

__temp__ = rimport("ggplot2")
using .__temp__

source

RCall.@rput — Macro.

Copies variables from Julia to R using the same name.

source

RCall.@var_str — Macro.

Returns a variable named "str". Useful for passing keyword arguments containing dots.

source

Constants

RCall.globalEnv — Constant.

R global Environment.

globalEnv[:x] = 1
globalEnv[:x]

source

RCall.jtypExtPtrs — Constant.

Julia types (typically functions) which are wrapped in Ptr{ExtPtrSxp} are stored here to prevent garbage collection by Julia.

source

RCall.typs — Constant.

vector of R Sxp types

source